Move mypyc.class_ir to mypyc.ir.class_ir
