+++ mypyc/emitclass.py
@@ -13,7 +13,7 @@ from mypyc.emitwrapper import (
 )
 from mypyc.rtypes import RType, RTuple, object_rprimitive
 from mypyc.ir.func_ir import FuncIR, FuncDecl, FUNC_STATICMETHOD, FUNC_CLASSMETHOD
-from mypyc.class_ir import ClassIR, VTableMethod, VTableEntries
+from mypyc.ir.class_ir import ClassIR, VTableMethod, VTableEntries
 from mypyc.sametype import is_same_type
 from mypyc.namegen import NameGenerator
