+++ mypyc/emitfunc.py
@@ -14,7 +14,7 @@ from mypyc.ir.ops import (
 )
 from mypyc.rtypes import RType, RTuple
 from mypyc.ir.func_ir import FuncIR, FuncDecl, FUNC_STATICMETHOD, FUNC_CLASSMETHOD
-from mypyc.class_ir import ClassIR
+from mypyc.ir.class_ir import ClassIR
 from mypyc.namegen import NameGenerator
 
 # Whether to insert debug asserts for all error handling, to quickly
