+++ mypyc/emitmodule.py
@@ -35,7 +35,7 @@ from mypyc.emitwrapper import (
 from mypyc.ir.ops import LiteralsMap, DeserMaps
 from mypyc.rtypes import RType, RTuple
 from mypyc.ir.func_ir import FuncIR
-from mypyc.class_ir import ClassIR
+from mypyc.ir.class_ir import ClassIR
 from mypyc.module_ir import ModuleIR, ModuleIRs, deserialize_modules
 from mypyc.options import CompilerOptions
 from mypyc.uninit import insert_uninit_checks
