+++ mypyc/ir/class_ir.py
