+++ mypyc/ir/ops.py
@@ -30,7 +30,7 @@ from mypyc.rtypes import (
 from mypyc.common import short_name
 
 if TYPE_CHECKING:
-    from mypyc.class_ir import ClassIR  # noqa
+    from mypyc.ir.class_ir import ClassIR  # noqa
     from mypyc.ir.func_ir import FuncIR, FuncDecl  # noqa
 
 T = TypeVar('T')
