+++ mypyc/irbuild/builder.py
@@ -42,7 +42,7 @@ from mypyc.rtypes import (
     none_rprimitive, is_none_rprimitive, object_rprimitive, is_object_rprimitive
 )
 from mypyc.ir.func_ir import FuncIR, FuncSignature, FuncDecl, INVALID_FUNC_DEF
-from mypyc.class_ir import ClassIR, NonExtClassInfo
+from mypyc.ir.class_ir import ClassIR, NonExtClassInfo
 from mypyc.primitives.registry import func_ops
 from mypyc.primitives.list_ops import (
     list_append_op, list_len_op, new_list_op, to_list, list_pop_last
