+++ mypyc/irbuild/classdef.py
@@ -13,7 +13,7 @@ from mypyc.rtypes import (
     is_object_rprimitive, is_none_rprimitive
 )
 from mypyc.ir.func_ir import FuncIR, FuncDecl, FuncSignature, RuntimeArg
-from mypyc.class_ir import ClassIR, NonExtClassInfo
+from mypyc.ir.class_ir import ClassIR, NonExtClassInfo
 from mypyc.primitives.misc_ops import (
     dataclass_sleight_of_hand, py_setattr_op, pytype_from_template_op, py_calc_meta_op,
     type_object_op, py_hasattr_op, not_implemented_op, true_op
