+++ mypyc/test/test_serialization.py
@@ -10,7 +10,7 @@ from collections.abc import Iterable
 from mypyc.ir.ops import DeserMaps
 from mypyc.rtypes import RType
 from mypyc.ir.func_ir import FuncDecl, FuncIR, FuncSignature
-from mypyc.class_ir import ClassIR
+from mypyc.ir.class_ir import ClassIR
 from mypyc.module_ir import ModuleIR, deserialize_modules
 from mypyc.sametype import is_same_type, is_same_signature
