+++ mypyc/test/test_tuplename.py
@@ -4,7 +4,7 @@ from mypyc.rtypes import (
     RTuple, object_rprimitive, int_rprimitive, bool_rprimitive, list_rprimitive,
     RInstance, RUnion,
 )
-from mypyc.class_ir import ClassIR
+from mypyc.ir.class_ir import ClassIR
 
 
 class TestTupleNames(unittest.TestCase):
